pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adi-decision"
version = "0.1.0"
description = "ADI — Agent Decision Intelligence: structured multi-criteria decisions with policy, explainability, and confidence. Use as a library or via OpenAI/Anthropic agent tools."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["decision-making", "mcda", "multi-criteria", "explainability", "agent-tools", "llm-tools"]
authors = [
    { name = "ADI Contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.6.0",
    "numpy>=1.26.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
cli = [
    "typer>=0.9.0",
]
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
]
all = [
    "typer>=0.9.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
]

[project.urls]
Homepage = "https://github.com/dimgouso/adi-decision"
Repository = "https://github.com/dimgouso/adi-decision"
Issues = "https://github.com/dimgouso/adi-decision/issues"

[project.scripts]
adi = "adi.interfaces.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["adi"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--tb=short -q"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
